Transition to `f32` and `Framed` types, `Angle`, `Circle`, and `Direction` convergence

+++ crates/step_planning/src/traits/length.rs
@@ -1,17 +1,17 @@
 use crate::geometry::{Arc, LineSegment, Path, PathSegment};
 
 pub trait Length {
-    fn length(&self) -> f64;
+    fn length(&self) -> f32;
 }
 
 impl Length for Path {
-    fn length(&self) -> f64 {
+    fn length(&self) -> f32 {
         self.segments.iter().map(PathSegment::length).sum()
     }
 }
 
 impl Length for PathSegment {
-    fn length(&self) -> f64 {
+    fn length(&self) -> f32 {
         match self {
             PathSegment::LineSegment(line_segment) => line_segment.length(),
             PathSegment::Arc(arc) => arc.length(),
@@ -20,7 +20,7 @@ impl Length for PathSegment {
 }
 
 impl Length for LineSegment {
-    fn length(&self) -> f64 {
+    fn length(&self) -> f32 {
         let Self(start, end) = self;
 
         (*end - *start).norm()
@@ -28,7 +28,7 @@ impl Length for LineSegment {
 }
 
 impl Length for Arc {
-    fn length(&self) -> f64 {
+    fn length(&self) -> f32 {
         let angle_start_to_end = self.start.angle_to(self.end, self.direction);
 
         self.circle.radius * angle_start_to_end.into_inner()
